created addCharacterskillmethod

+++ Services/CharacterService/CharacterService.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
 using dotnet_rpg.Data;
@@ -11,35 +12,44 @@ namespace dotnet_rpg.Services.CharacterService
 {
     public class CharacterService : ICharacterService
     {
-       
+
         private readonly IMapper _mapper;
-
         private readonly DataContext _context;
+        private readonly IHttpContextAccessor _httpContextAccessor;
 
-        public CharacterService(IMapper mapper, DataContext context)
+        public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
         {
             _mapper = mapper;
             _context = context;
+            _httpContextAccessor = httpContextAccessor;
         }
+
+        private int GetUserId() => int.Parse(_httpContextAccessor.HttpContext.User
+                    .FindFirstValue(ClaimTypes.NameIdentifier));
+
+
         public async Task<ServiceResponse<List<GetCharacterDto>>> AddCharacter(AddCharacterDto newCharacter)
         {
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
             Character character = _mapper.Map<Character>(newCharacter);
 
-            _context.Characters.Add(character);
+            character.User = await _context.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
+
+            await _context.Characters.AddAsync(character);
             await _context.SaveChangesAsync();                          //gen new id for char
             serviceResponse.Data = await _context.Characters
+                    .Where(c => c.User.Id == GetUserId())
                     .Select(c => _mapper.Map<GetCharacterDto>(c))
                     .ToListAsync();
             return serviceResponse;
         }
 
-        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters(int userId)
+        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
         {
             var response = new ServiceResponse<List<GetCharacterDto>>();
 
             var dbCharacters = await _context.Characters
-            .Where( c => c.User.Id == userId )
+            .Where(c => c.User.Id == GetUserId())
             .ToListAsync();
 
             response.Data = dbCharacters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
@@ -50,7 +60,7 @@ namespace dotnet_rpg.Services.CharacterService
         public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
         {
             var serviceResponse = new ServiceResponse<GetCharacterDto>();
-            var DBcharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id);
+            var DBcharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
             serviceResponse.Data = _mapper.Map<GetCharacterDto>(DBcharacter);
             return serviceResponse;
         }
@@ -61,7 +71,6 @@ namespace dotnet_rpg.Services.CharacterService
 
             try
             {
-
 
                 var character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == updateCharacter.Id);
                 _mapper.Map(updateCharacter, character);
@@ -88,8 +97,6 @@ namespace dotnet_rpg.Services.CharacterService
 
             return response;
 
-
-
         }
 
         public async Task<ServiceResponse<List<GetCharacterDto>>> DeleteCharacter(int id)
@@ -99,11 +106,23 @@ namespace dotnet_rpg.Services.CharacterService
             try
             {
 
-                var character = await _context.Characters.FirstAsync(c => c.Id == id);
-                _context.Characters.Remove(character);
-                await _context.SaveChangesAsync();
+                var character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
+                if (character != null)
+                {
+                    _context.Characters.Remove(character);
+                    await _context.SaveChangesAsync();
 
-                response.Data = await  _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
+                    response.Data = await _context.Characters
+                        .Where( c => c.User.Id == GetUserId())
+                        .Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
+                }
+                else
+                {
+                    response.Success = false;
+                    response.Message = "Character not found";
+                }
+
+
 
 
             }
@@ -116,13 +135,48 @@ namespace dotnet_rpg.Services.CharacterService
 
             return response;
 
-
-
         }
 
+        public async Task<ServiceResponse<GetCharacterDto>> AddCharacterSkill(AddCharacterSkillDto newCharacterSkill)
+        {
+            var response = new ServiceResponse<GetCharacterDto>();
+
+            try
+            {
+                var character = await _context.Characters
+                .Include( c=> c.Weapon)
+                .Include( c=> c.Skills)
+                .FirstOrDefaultAsync( c=> c.Id == newCharacterSkill.CharacterId && c.User.Id == GetUserId());
+
+                if (character == null)
+                {
+                    response.Success = false;
+                    response.Message = "Character not found.";
+                    return response;
+                }
+
+                var skill = await _context.Skills.FirstOrDefaultAsync( s=> s.Id == newCharacterSkill.SkillId);
+                if (skill == null)
+                {
+                    response.Success = false;
+                    response.Message = "Skill not found.";
+                    return response;
+                }
+
+                character.Skills.Add(skill);
+
+                response.Data = _mapper.Map<GetCharacterDto>(character);
 
 
+            }
+            catch (System.Exception ex)
+            {
+                
+                response.Success = false;
+                response.Message = ex.Message;
+            }
 
-
+            return response;
+        }
     }
 }
